feat: port prose stat-instruction extraction to templates (shared module)
The family prose machinery (PR #123) moves to pfsrd2/prose_changes.py,
shared by both parsers. The template parser scans the intro prose of
ul-bearing creation sections — gated there because that's where
instructions live and flavor sections must not match. Document order:
intro prose precedes the <li> changes.

Recovered: the level+1 change missing from ALL FOUR Dark Archive
cryptid templates and Broodpiercer ('Increase the creature's level by 1
and change its statistics as follows.' lived only in intro prose —
user-reported on Experimental Cryptid, repair epic bd_521Studios-6xxc).
Zero false positives, zero removals across the corpus; family output
byte-identical.
